Added new content layout component
New component, 'ContentLayout', has been added in the project under 'src/app/layout'. It includes HTML, SCSS, and spec files. This addition for having new content layout for the design changes needed.

The component contains a header and a main section. The header includes logo and a call-to-action. The main section carries other site content (logo, heading, content, cta) and a floating button for immediate contact or call.

This design provides clarity for users on where to find key information and how they can take the next step. The component allows easier maintenance and future improvements of this primary section of the site.